🌈 add feedback after mark as done a task
