+++ src/utils/constants.js
@@ -3,7 +3,7 @@ module.exports = {
   // add
   ADD_EMPTY_ITEM: 'You forgot the item description',
   SUCESSFULLY_ADDED: item =>
-    `✅ Successfully added '${item}' in your todo list`,
+    `📆 Successfully added '${item}' in your todo list`,
   // clean
   CLEAN_EMPTY_LIST: 'Which list do you want to clean?',
   CLEAN_PROMPT_MSG: (is_all_lists, list) =>
@@ -13,8 +13,10 @@ module.exports = {
   // done
   DONE_INCORRECT_INDEX: index =>
     `⚠️  Does not exists an item with ID '${index}'`,
+  DONE_HAS_BEEN_MARKED: index => 
+    `✅ Item ${index} has been marked as done`,
   // list
   LIST_EMPTY: '📭 Empty lists',
-  LIST_TODOS: todos => `📆 Todo\n${todos()}`,
-  LIST_DONES: dones => `✅ Done\n${dones()}`,
+  LIST_TODOS: todos => `📆 Todo\n${todos}`,
+  LIST_DONES: dones => `✅ Done\n${dones}`,
 };
